Update to wgpu-matrix v3
This removes much of the casting that was needed before.

+++ sample/cameras/main.ts
@@ -196,7 +196,7 @@ function getModelViewProjectionMatrix(deltaTime: number) {
   const camera = cameras[params.type];
   const viewMatrix = camera.update(deltaTime, inputHandler());
   mat4.multiply(projectionMatrix, viewMatrix, modelViewProjectionMatrix);
-  return modelViewProjectionMatrix as Float32Array;
+  return modelViewProjectionMatrix;
 }
 
 let lastFrameMS = Date.now();
