+++ sample/samplerParameters/main.ts
@@ -276,7 +276,7 @@ const viewProj = mat4.translate(
   mat4.perspective(2 * Math.atan(1 / kCameraDist), 1, 0.1, 100),
   [0, 0, -kCameraDist]
 );
-device.queue.writeBuffer(bufConfig, 0, viewProj as Float32Array);
+device.queue.writeBuffer(bufConfig, 0, viewProj);
 
 const bufMatrices = device.createBuffer({
   usage: GPUBufferUsage.STORAGE,
